+++ TvDemo/src/PResoluciones/Resolucion4K.java
@@ -5,8 +5,8 @@ package PResoluciones;
  * Class ResolucionHD
  */
 public class Resolucion4K extends Resolucion {
-    protected final static int LARGO = 2160;
-    protected final static int ANCHO = 3840;
+    protected final static double LARGO = 2160;
+    protected final static double ANCHO = 3840;
     
     public Resolucion4K(){
         super(LARGO,ANCHO);
